pyproject.toml
[project]
name = "sync-resume"
version = "0.1.0"
description = "A robust resume parser that extracts structured information from resumes and stores the data in both traditional databases and vector stores. With SyncResume, candidates’ skills, experience, and qualifications are indexed to quickly connect them to the most relevant job opportunities."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "langchain-core>=0.3.35",
    "langchain-community>=0.3.17",
    "langchain-openai>=0.3.6",
    "langchain-postgres>=0.0.13",
    "langgraph>=0.2.73",
    "pillow>=11.1.0",
    "psycopg[binary]>=3.2.5",
    "pydantic[email]>=2.10.6",
    "pymupdf>=1.25.3",
    "python-dotenv>=1.0.1",
    "python-levenshtein>=0.26.1",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.6.1",
    "sqlalchemy[asyncio]>=2.0.38",
    "uvicorn>=0.34.0",
    "numpy>=1.26.4",
    "pandas>=2.2.3",
    "matplotlib>=3.10.1",
]

[dependency-groups]
dev = [
    "beautifulsoup4>=4.13.3",
    "fuzzywuzzy>=0.18.0",
    "ipykernel>=6.29.5",
    "openpyxl>=3.1.5",
    "pydantic-extra-types>=2.10.2",
    "pypdf>=5.3.0",
    "ruff>=0.9.6",
]
